requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.2
cycler==0.12.1
dotenv==0.9.9
fonttools==4.60.1
idna==3.11
joblib==1.5.2
kagglehub==0.3.13
kiwisolver==1.4.9
loguru==0.7.3
markdown-it-py==4.0.0
matplotlib==3.10.7
mdurl==0.1.2
numpy==1.26.4
opencv-python==4.10.0
opencv-python-headless==4.10.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pip==25.3
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
threadpoolctl==3.6.0
tqdm==4.67.1
typer==0.20.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wheel==0.45.1
